requirements.txt
flask 
web3 
requests 
python-dotenv
pycoingecko